requirements.txt
torch
transformers
datasets
peft
evaluate
scikit-learn
pandas
kagglehub
matplotlib
seaborn
accelerate
bitsandbytes
jupyterlab
ipywidgets
textstats
spacy